requirements.txt
black==22.1.0
click==8.0.4
flake8==4.0.1
mccabe==0.6.1
mwparserfromhell==0.6.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.9.0
platformdirs==2.5.0
protobuf==3.19.4
pycodestyle==2.8.0
pyflakes==2.4.0
toml==0.10.2
tomli==2.0.1
